+++ DRN.Framework.SharedKernel/Domain/Entity.cs
@@ -2,67 +2,42 @@ namespace DRN.Framework.SharedKernel.Domain;
 
 public abstract class Entity
 {
-    protected List<IDomainEvent> DomainEvents { get; } = new();
-
+    private List<IDomainEvent> DomainEvents { get; } = new();
     public IReadOnlyList<IDomainEvent> GetDomainEvents() => DomainEvents;
     public long Id { get; protected set; }
     public DateTimeOffset CreatedAt { get; protected set; }
     public DateTimeOffset ModifiedAt { get; protected set; }
-
-    public virtual void MarkAsCreated()
-    {
-        CreatedAt = DateTimeOffset.UtcNow;
-        ModifiedAt = DateTimeOffset.UtcNow;
-        var created = GetCreatedEvent();
-        AddDomainEvent(created);
-    }
-
-    public virtual void MarkAsModified()
-    {
-        ModifiedAt = DateTimeOffset.UtcNow;
-        var modified = GetModifiedEvent();
-        AddDomainEvent(modified);
-    }
-
-    public virtual void MarkAsDeleted()
-    {
-        var deleted = GetDeletedEvent();
-        AddDomainEvent(deleted);
-    }
-
-    public virtual EntityCreated? GetCreatedEvent() => null;
-
-    public virtual EntityModified? GetModifiedEvent() => null;
-
-    public virtual EntityDeleted? GetDeletedEvent() => null;
 
     protected void AddDomainEvent(DomainEvent? e)
     {
         if (e != null) DomainEvents.Add(e);
     }
 
-    protected bool Equals(Entity other)
+    public void MarkAsCreated()
     {
-        return Id == other.Id;
+        CreatedAt = DateTimeOffset.UtcNow;
+        ModifiedAt = DateTimeOffset.UtcNow;
+        AddDomainEvent(GetCreatedEvent());
     }
 
-    public override bool Equals(object? obj)
+    public void MarkAsModified()
     {
-        return ReferenceEquals(this, obj) || obj is Entity other && Equals(other);
+        ModifiedAt = DateTimeOffset.UtcNow;
+        AddDomainEvent(GetModifiedEvent());
     }
 
-    public override int GetHashCode()
+    public void MarkAsDeleted()
     {
-        return Id.GetHashCode();
+        AddDomainEvent(GetDeletedEvent());
     }
 
-    public static bool operator ==(Entity? left, Entity? right)
-    {
-        return Equals(left, right);
-    }
+    protected abstract EntityCreated? GetCreatedEvent();
+    protected abstract EntityModified? GetModifiedEvent();
+    protected abstract EntityDeleted? GetDeletedEvent();
 
-    public static bool operator !=(Entity? left, Entity? right)
-    {
-        return !Equals(left, right);
-    }
+    private bool Equals(Entity other) => Id == other.Id;
+    public override bool Equals(object? obj) => ReferenceEquals(this, obj) || obj is Entity other && Equals(other);
+    public override int GetHashCode() => Id.GetHashCode();
+    public static bool operator ==(Entity? left, Entity? right) => Equals(left, right);
+    public static bool operator !=(Entity? left, Entity? right) => !Equals(left, right);
 }