+++ DRN.Framework.SharedKernel/Exceptions.cs
@@ -1,66 +1,16 @@
 namespace DRN.Framework.SharedKernel;
 
-public abstract class DrnException : Exception
+public abstract class DrnException(string message, Exception exception = null!, string? category = "default") : Exception(message, exception)
 {
-    public string Category { get; }
-
-    protected DrnException(string message, Exception exception = null!, string category = "default") : base(message, exception)
-    {
-        Category = category;
-    }
+    public string Category { get; } = category ?? "default";
 }
 
-public class ValidationException : DrnException
-{
-    public ValidationException(string message, string category, Exception exception = null!) : base(message, exception, category)
-    {
-    }
+public class ValidationException(string message, Exception exception = null!, string? category = null) : DrnException(message, exception, category);
 
-    public ValidationException(string message, Exception exception = null!) : base(message, exception)
-    {
-    }
-}
+public class NotSavedException(string message, Exception exception = null!, string? category = null) : DrnException(message, exception, category);
 
-public class NotSavedException : DrnException
-{
-    public NotSavedException(string message, string category, Exception exception = null!) : base(message, exception, category)
-    {
-    }
+public class NotFoundException(string message, Exception exception = null!, string? category = null) : DrnException(message, exception, category);
 
-    public NotSavedException(string message, Exception exception = null!) : base(message, exception)
-    {
-    }
-}
+public class ExpiredException(string message, Exception exception = null!, string? category = null) : DrnException(message, exception, category);
 
-public class NotFoundException : DrnException
-{
-    public NotFoundException(string message, string category, Exception exception = null!) : base(message, exception, category)
-    {
-    }
-
-    public NotFoundException(string message, Exception exception = null!) : base(message, exception)
-    {
-    }
-}
-
-public class ExpiredException : DrnException
-{
-    public ExpiredException(string message, string category, Exception exception = null!) : base(message, exception, category)
-    {
-    }
-
-    public ExpiredException(string message, Exception exception = null!) : base(message, exception)
-    {
-    }
-}
-
-public class ConfigurationException : DrnException
-{
-    public ConfigurationException(string message, string category, Exception exception = null!) : base(message, exception, category)
-    {
-    }
-
-    public ConfigurationException(string message, Exception exception = null!) : base(message, exception)
-    {
-    }
-}
+public class ConfigurationException(string message, Exception exception = null!, string? category = null) : DrnException(message, exception, category);