+++ Sample.Domain/QA/Answers/AnswerComment.cs
@@ -20,4 +20,9 @@ public class AnswerComment : AggregateRoot
     public long AnswerId { get; private set; }
 
     public List<AnswerComment> Comments { get; private set; } = [];
+
+    protected override EntityCreated? GetCreatedEvent() => null;
+    protected override EntityModified? GetModifiedEvent() => null;
+    protected override EntityDeleted? GetDeletedEvent() => null;
+
 }